% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloques.R
\name{bloques}
\alias{bloques}
\title{bloques ar 1}
\usage{
bloques(s, r, t, n)
}
\arguments{
\item{s}{Numerical value indicating global standard deviation of the matrix}

\item{r}{Numerical value indicating correlation of individuals}

\item{t}{Numerical value indicating number of times when observations are repeated}

\item{n}{Numerical value indicating number of individuals}
}
\value{
A diagonal block matrix with structure AR(1)
}
\description{
Build a block diagonal matrix with structure AR(1)
}
\examples{
bloques(2,.5,10,2)
}
\references{
Nuñez A. and Zimmerman D. 2001. Modelación de datos longitudinales con estructuras de covarianza no estacionarias: Modelo de coeficientes aleatorios frente a modelos alternativos. Questio. 2001. 25.
}
