% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloques.R
\name{bloques2}
\alias{bloques2}
\title{bloques arma (1,1)}
\usage{
bloques2(s, r, g, t, n)
}
\arguments{
\item{s}{Numerical value indicating global standard deviation of the matrix}

\item{r}{Numerical value indicating the first parameter rho correlation of individuals}

\item{g}{Numerical value indicating the second parameter phi correlation of individuals}

\item{t}{Numerical value indicating number of times when observations are repeated}

\item{n}{Numerical value indicating number of individuals}
}
\value{
A diagonal block matrix with structure ARMA(1,1)
}
\description{
Build a block diagonal matrix with structure ARMA(1,1)
}
\examples{
bloques2(2,.5,.8,10,2)
}
\references{
Nuñez A. and Zimmerman D. 2001. Modelación de datos longitudinales con estructuras de covarianza no estacionarias: Modelo de coeficientes aleatorios frente a modelos alternativos. Questio. 2001. 25.
}
