% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\docType{data}
\name{Dental}
\alias{Dental}
\title{Dental distance}
\format{A data Frame with 98 rows and 5 variables:
\describe{
  \item{gencode}{1 for girls, 0 for boy}
  \item{id}{Number of the individual}
  \item{distance}{Distance from the center of the pituitary gland to the pterygomaxillary fissure}
  \item{age}{Child's age at which measurement was taken}
  \item{gender}{Gender of the child}
 }}
\source{
\url{https://faculty.biostat.ucla.edu/robweiss/filedepot_download/87/524}
}
\usage{
Dental
}
\description{
It reports the distance in millimeters from the center of the pituitary to the pteromaxillary fissure. The subjects were 16 children and 11 girls. Data were taken every two years from 8 years and ended at age 14.
}
\keyword{datasets}
